+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.2.0
-
-Initial release.
+- Initial release.
 
 ## 0.2.1
+- Drop unnecessary ORDER BY columns following a non-nullable and distinct column.
 
-Drop unnecessary ORDER BY columns following a non-nullable and distinct column.
+## 0.2.2
+- Raise error when there is no non-nullable and distinct column in the configured order definition.